+++ hipaycard-cmp/src/iosMain/kotlin/com/hipay/card/cmp/CurrentYear.ios.kt
@@ -1,0 +1,8 @@
+package com.hipay.card.cmp
+
+import platform.Foundation.NSCalendar
+import platform.Foundation.NSCalendarUnitYear
+import platform.Foundation.NSDate
+
+internal actual fun currentYear(): Int =
+    NSCalendar.currentCalendar.components(NSCalendarUnitYear, fromDate = NSDate()).year.toInt()